youtube_transcript/requirements.txt
streamlit
google-generativeai 
python-dotenv 
pdf2image 
youtube-transcript-api

